partition_snapshot_reader: Move range tombstone stream member
The lsa_partition_reader is the helper sub-class for
partition_snapshot_reader that, among other things, is
responsible for filling the stream of range tombstones,
that's then used by the reader itself.

Next patches will change the way range tombstones are
emitted by the reader, so hide the stream inside the
helper subclass in advance.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>