cdc: switch to the new generation management method
When we notice a new generation timestamp in gossip we use an RPC call
to fetch the generation data instead of a distributed table.

We then store the data in a local table before we start gossiping it as
well so we can also answer fetch requests from other nodes.