transport: Update system.clients in background
Also, log exceptions specifically thrown on these updates.

The change is introduced for more resilience and easier debuggability
(i.e. not motivated by a particular bug). If the update/delete to
`system.clients` hangs for any reason (e.g. overload) we should not
wait for it - just process the connection as if nothing happened.
If connection processing also throws, we'll get separate exceptions
instead of a nested exception.

Refs: scylladb/scylla-enterprise#1618